requirements.txt
fastapi
uvicorn
python-telegram-bot[asyncio]
sqlalchemy
asyncpg
alembic
pydantic
python-dotenv
celery
redis
httpx
matplotlib
openai  # optional for AI tips
pytest
pytest-asyncio
python-dotenv
beautifulsoup4
requests